// src/config/llm-presets/types.ts
// Copyright 2025-2026 miaotouy(Github@miaotouy)
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

/**
 * 预设模板接口
 */

import type {
  ProviderType,
  LlmModelInfo,
  LlmLink,
  LlmProfile,
} from "../../types/llm-profiles";

export interface LlmPreset {
  type: ProviderType;
  name: string;
  description: string;
  defaultBaseUrl: string;
  logoUrl?: string;
  defaultModels?: LlmModelInfo[];
  links?: LlmLink[];
  /** 高级端点配置预设 */
  customEndpoints?: LlmProfile["customEndpoints"];
}